requirements.txt
langchain_community
langchain_core
langchain_upstage
langchain-chroma
langchain-google-genai
python-dotenv
SQLAlchemy
langgraph
chromadb
streamlit
pytest
sqlalchemy-libsql